Add ability to copy only local files, using the -l switch

+++ build.py
@@ -17,11 +17,20 @@ from compressions.zip   import ZipArchive
 from compressions.tar   import TarArchive
 
 # Ensure paths are set correctly
+args = sys.argv[1:]
+
+# Check if the -l switch is set
+if "-l" in args:
+    local_only = True
+    args.remove('-l')
+else:
+    local_only = False
+
 try:
-    if len(sys.argv[1:]) == 2:
-        config_path, project_path = sys.argv[1:]
-    elif len(sys.argv[1:]) == 1:
-        config_path = sys.argv[1:2][0]
+    if len(args) == 2:
+        config_path, project_path = args
+    elif len(args) == 1:
+        config_path = args[0]
         project_path = None
 except ValueError:
     print 'build.py requires at least a config path'
@@ -32,7 +41,7 @@ print config_path
 # Load config XML file
 config = parse(config_path)
 
-# Set relevent paths based on configeration data
+# Set relevent paths based on configuration data
 wc_location = config.getElementsByTagName('workingcopy')[0].getAttribute('location')
 wc_path = os.path.dirname(config_path) + os.sep
     
@@ -52,61 +61,63 @@ if not project_path.endswith(os.sep):
     project_path += os.sep
     
 if not os.path.exists(project_path):
-    os.mkdir(project_path)
+    os.makedirs(project_path)
     print "Directory Created: " + project_path
 
-# Generate a list of ignores from he config file
+# Generate a list of ignores from the config file
 ignores = []
 for ignore in config.getElementsByTagName('ignore'):
     ignores.append(ignore.getAttribute('name'))
 
-# Generate a list of dicts to store the XML remotes data
-remotes = []
 
-for remote in config.getElementsByTagName('remote'):
-    remotes.append({
-        'name': remote.getAttribute('name'),
-        'url': remote.getAttribute('url'),
-        'type': remote.getAttribute('type')
-    })
-
-# Loop through each remote, download, extract then delete it
-for item in remotes:
-    filename = item['name'] + '.' + item['type']
-    download_file = project_path + filename
-    extract_path = project_path + item['name']
-        
-    # Check if remote has been downloaded before, if so, remove it
-    if os.path.isdir(extract_path):
-        shutil.rmtree(extract_path)
-        print "Removed module: " + item['name']
+if not local_only:
+    # Generate a list of dicts to store the XML remotes data
+    remotes = []
     
-    print 'Downloading: ' + filename
-    remote = urllib.urlopen(item['url'])
-    local = open(download_file, 'w')
-    local.write(remote.read())
+    for remote in config.getElementsByTagName('remote'):
+        remotes.append({
+            'name': remote.getAttribute('name'),
+            'url': remote.getAttribute('url'),
+            'type': remote.getAttribute('type')
+        })
+    
+    # Loop through each remote, download, extract then delete it
+    for item in remotes:
+        filename = item['name'] + '.' + item['type']
+        download_file = project_path + filename
+        extract_path = project_path + item['name']
             
-    print 'Downloaded: ' + filename
-    remote.close()
-    local.close()
-          
-    if item['type'] == "zip":
-        archive = ZipArchive(download_file)
-          
-    elif item['type'] == "tar":
-        archive = TarArchive(download_file)
-          
-    elif item['type'] == "tar.gz":
-        archive = TarArchive(download_file,'gz')
-          
-    elif item['type'] == "tar.bz2":
-        archive = TarArchive(download_file,'bz2')
+        # Check if remote has been downloaded before, if so, remove it
+        if os.path.isdir(extract_path):
+            shutil.rmtree(extract_path)
+            print "Removed module: " + item['name']
         
-    if archive:
-        archive.extract(extract_path)
+        print 'Downloading: ' + filename
+        remote = urllib.urlopen(item['url'])
+        local = open(download_file, 'w')
+        local.write(remote.read())
                 
-    os.remove(download_file)
-    print "Removed: " + filename
+        print 'Downloaded: ' + filename
+        remote.close()
+        local.close()
+              
+        if item['type'] == "zip":
+            archive = ZipArchive(download_file)
+              
+        elif item['type'] == "tar":
+            archive = TarArchive(download_file)
+              
+        elif item['type'] == "tar.gz":
+            archive = TarArchive(download_file,'gz')
+              
+        elif item['type'] == "tar.bz2":
+            archive = TarArchive(download_file,'bz2')
+            
+        if archive:
+            archive.extract(extract_path)
+                    
+        os.remove(download_file)
+        print "Removed: " + filename
     
 if not wc_path == project_path:
     # Loop through all files and directories in working copy them to project dir
